refactor(cli-session): move tests next to register_and_poll and plan_pr_drift
Audit follow-up (2026-05-12):
- register_and_poll.rs sat 4 lines below the 300-line cap.
- check_1_plan_pr_drift.rs mixed shelling, parsing, http, and tests.

Tests now live in sibling files via the #[path] / mod tests pattern
already used by session_tests.rs.
